Completed visual block mode deletion, cc command to go into insert after line-deletion, comma to reverse direction of ; find command, and auto closing pairs

+++ src/Buffer.cpp
@@ -203,7 +203,7 @@ int Buffer::findCharacterIndex(char character, bool findForwards)
     }
     else
     {
-        for (size_t i = m_cursorX - 1; i > 0; i--)
+        for (int i = m_cursorX - 1; i > 0; i--)
         {
             if (lineGapBuffer->at(i) == character)
             {
